Improvements to the ImageDiff node and updated demonstration workflow
The ImageDiff node now has a pair of toggles: when console_mse is false the Mean Square Error calculation is skipped; when contrast_boost is true the difference between images is boosted by a linear scaling factor (currently set to 5).

The calculation of difference has also been modified to take the absolute value, the resulting image is now purely magnitudinal and does not include directionality of change (I think this should pretty much always be better suited to the imagined use case).